Rename errors.py to exceptions.py
Be consistent with our other packages.
